# backend/docs/plan/task-result-notifications/gateway-notification-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LazyMind Channel Gateway",
    "description": "Unified external chat channel gateway.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/channel-gateway/v1/notification-capabilities": {
      "get": {
        "summary": "Notification Capabilities",
        "operationId": "notification_capabilities_api_channel_gateway_v1_notification_capabilities_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationCapabilitiesView"
                }
              }
            }
          },
          "400": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "401": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "403": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "404": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "409": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "422": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "500": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "503": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          }
        }
      }
    },
    "/api/channel-gateway/v1/channel-accounts/{account_id}/notification-recipients": {
      "get": {
        "summary": "Notification Recipients",
        "operationId": "notification_recipients_api_channel_gateway_v1_channel_accounts__account_id__notification_recipients_get",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          },
          {
            "name": "page_token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 2048,
              "default": "",
              "title": "Page Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationRecipientsView"
                }
              }
            }
          },
          "400": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "401": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "403": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "404": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "409": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "422": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "500": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "503": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          }
        }
      }
    },
    "/api/channel-gateway/v1/channel-accounts/{account_id}/notification-references": {
      "get": {
        "summary": "Notification References",
        "operationId": "notification_references_api_channel_gateway_v1_channel_accounts__account_id__notification_references_get",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationReferencesView"
                }
              }
            }
          },
          "400": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "401": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "403": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "404": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "409": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "422": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "500": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "503": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          }
        }
      }
    },
    "/api/channel-gateway/v1/channel-accounts/{account_id}/disconnect-impact": {
      "get": {
        "summary": "Notification Disconnect Impact",
        "operationId": "notification_disconnect_impact_api_channel_gateway_v1_channel_accounts__account_id__disconnect_impact_get",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationReferencesView"
                }
              }
            }
          },
          "400": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "401": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "403": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "404": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "409": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "422": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "500": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "503": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          }
        }
      }
    },
    "/api/channel-gateway/v1/channel-accounts/{account_id}:reconnect": {
      "post": {
        "summary": "Notification Reconnect",
        "operationId": "notification_reconnect_api_channel_gateway_v1_channel_accounts__account_id__reconnect_post",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionSessionView"
                }
              }
            }
          },
          "400": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "401": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "403": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "404": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "409": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "422": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "500": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          },
          "503": {
            "description": "Safe notification business error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationErrorView"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConnectionSessionView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "mode": {
            "type": "string",
            "const": "qr_code",
            "title": "Mode"
          },
          "status": {
            "type": "string",
            "enum": [
              "preparing",
              "waiting_scan",
              "scanned",
              "verification_required",
              "confirming",
              "connected",
              "expired",
              "canceled",
              "failed"
            ],
            "title": "Status"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "qr": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QRCodeView"
              },
              {
                "type": "null"
              }
            ]
          },
          "challenge": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChallengeView"
              },
              {
                "type": "null"
              }
            ]
          },
          "poll_after_ms": {
            "type": "integer",
            "title": "Poll After Ms"
          },
          "allowed_actions": {
            "items": {
              "type": "string",
              "enum": [
                "cancel",
                "submit_challenge",
                "refresh"
              ]
            },
            "type": "array",
            "title": "Allowed Actions"
          },
          "account": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccountView"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SessionErrorView"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "provider",
          "mode",
          "status",
          "revision",
          "message",
          "qr",
          "challenge",
          "poll_after_ms",
          "allowed_actions",
          "account",
          "error"
        ],
        "title": "ConnectionSessionView"
      },
...
      "NotificationCapabilitiesView": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NotificationCapabilityView"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "NotificationCapabilitiesView"
      },
...
      "NotificationErrorView": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/NotificationErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "NotificationErrorView"
      },
...
      "NotificationRecipientsView": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NotificationRecipientView"
            },
            "type": "array",
            "title": "Items"
          },
          "next_page_token": {
            "type": "string",
            "title": "Next Page Token"
          },
          "verified_person": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NotificationRecipientView"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_page_token"
        ],
        "title": "NotificationRecipientsView"
      },
...
      "NotificationReferencesView": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NotificationReferenceView"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "NotificationReferencesView"
      },
...
    }
  }
}